test: FAQ 리스트 삭제 - 빈 리스트 전달되는 케이스 추가

+++ src/main/java/com/project/swimcb/faq/application/DeleteFaqsInteractor.java
@@ -17,6 +17,9 @@ public class DeleteFaqsInteractor implements DeleteFaqsUseCase {
 
   @Override
   public void deleteAll(@NonNull List<Long> faqIds) {
+    if (faqIds.isEmpty()) {
+      return;
+    }
     faqRepository.deleteAllById(faqIds);
   }
 }
